libft test

+++ libft/srcs/ft_atoi.c
@@ -27,7 +27,7 @@ int ft_atoi(const char *str)
 {
     size_t i;
     int sign;
-    long result;
+    long long result;
 
     i = 0;
     sign = 1;
